serial_tvbox_python/requirements.txt
antiorm==1.2.1
db==0.1.1
db-sqlite3==0.0.1
pyserial==3.5
